LPD-33439 Moved asset display page friendly url resolver to the right package
