+++ modules/apps/asset/asset-categories-admin-web/src/main/java/com/liferay/asset/categories/admin/web/internal/info/display/page/portlet/AssetCategoryAssetDisplayPageFriendlyURLResolver.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.asset.categories.admin.web.internal.info.display.contributor.portlet;
+package com.liferay.asset.categories.admin.web.internal.info.display.page.portlet;
 
 import com.liferay.asset.display.page.portlet.BaseAssetDisplayPageFriendlyURLResolver;
 import com.liferay.asset.kernel.model.AssetCategory;
