+++ modules/apps/calendar/calendar-web/src/main/java/com/liferay/calendar/web/internal/asset/display/page/portlet/CalendarBookingAssetDisplayPageFriendlyURLResolver.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.calendar.web.internal.asset.display.contributor.portlet;
+package com.liferay.calendar.web.internal.asset.display.page.portlet;
 
 import com.liferay.asset.display.page.portlet.BaseAssetDisplayPageFriendlyURLResolver;
 import com.liferay.portal.kernel.portlet.FriendlyURLResolver;
